requirements.txt
anyio==4.12.1
asgiref==3.11.1
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
Django==6.0.3
django-filter==25.2
djangorestframework==3.16.1
firebase_admin==7.2.0
google-api-core==2.30.0
google-auth==2.49.0
google-cloud-core==2.5.0
google-cloud-firestore==2.24.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Markdown==3.10.2
msgpack==1.1.2
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.2
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
